Modularity by creation of libraries
