+++ TelCo.ColorCoder.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.852
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30413.136
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TelCo.ColorCoder", "TelCo.ColorCoder\TelCo.ColorCoder.csproj", "{7620DC68-E947-49D3-B846-44666F3FB907}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TelCo.ColorCoder", "TelCo.ColorCoder\TelCo.ColorCoder.csproj", "{7620DC68-E947-49D3-B846-44666F3FB907}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
